Android: Bugfixes

+++ clients/android/NewsBlur/src/com/newsblur/util/ImageCache.java
@@ -48,6 +48,7 @@ public class ImageCache {
             }
 
             File f = new File(cacheDir, fileName);
+            if (f.exists()) return;
             URL u = new URL(url);
             NetworkUtils.loadURL(u, new FileOutputStream(f));
         } catch (IOException e) {
